requirements.txt
python-telegram-bot==20.7
httpx==0.27.0
gspread==6.0.2
google-auth==2.29.0
python-dotenv==1.0.1
